pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "learning-assistant"
version = "0.2.0"
description = "A modular, plugin-based AI learning assistant CLI tool with Agent integration support"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
keywords = ["ai", "learning", "video", "summary", "cli"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # Core Framework
    "typer>=0.9.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "loguru>=0.7.0",

    # LLM Integration (Official SDKs only)
    "openai>=1.0.0",
    "anthropic>=0.18.0",

    # Video Processing
    "yt-dlp",
    "faster-whisper>=0.10.0",
    "yutto>=2.2.0",  # B站-specific downloader (avoids SSL issues)

    # Authentication
    "qrcode>=7.4.0",

    # Web Processing
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "trafilatura>=1.6.0",
    "aiohttp>=3.9.0",
    "python-dateutil>=2.8.0",

    # Data Processing
    "jinja2>=3.1.0",
    "orjson>=3.9.0",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",

    # Code Quality
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",

    # Documentation
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.0.0",
]

link = [
    # Optional dependencies for Link Learning module
    "playwright>=1.40.0",
    "readability-lxml>=0.8.1",
]

[project.scripts]
la = "learning_assistant.cli:app"

[project.urls]
Homepage = "https://github.com/yourname/learning-assistant"
Documentation = "https://learning-assistant.readthedocs.io"
Repository = "https://github.com/yourname/learning-assistant.git"
Issues = "https://github.com/yourname/learning-assistant/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/learning_assistant"]

[tool.black]
line-length = 88
target-version = ['py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # defaults
  \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  # project specific
  | data
  | cache
)/
'''

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "B027",  # empty method in abstract base class (optional methods)
]

[tool.ruff.lint.isort]
known-first-party = ["learning_assistant"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
strict_optional = true

[[tool.mypy.overrides]]
module = [
    "yt_dlp",
    "whisper",
    "trafilatura",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short --strict-markers"
markers = [
    "slow: marks tests as slow",
    "integration: marks tests as integration tests",
]